Generalize EditableListFilter
