+++ indico/modules/events/editing/client/js/management/editable_type/EditableList.jsx
@@ -19,12 +19,13 @@ import PropTypes from 'prop-types';
 import React, {useState, useMemo} from 'react';
 import {useParams, Link} from 'react-router-dom';
 import {Column, Table, SortDirection, WindowScroller} from 'react-virtualized';
-import {Button, Icon, Input, Loader, Checkbox, Message, Dropdown, Label} from 'semantic-ui-react';
+import {Button, Icon, Loader, Checkbox, Message, Dropdown} from 'semantic-ui-react';
 
 import {
   TooltipIfTruncated,
   ManagementPageSubTitle,
   ManagementPageBackButton,
+  ListFilter,
 } from 'indico/react/components';
 import {useIndicoAxios} from 'indico/react/hooks';
 import {Translate} from 'indico/react/i18n';
@@ -105,7 +106,6 @@ function EditableListDisplay({
   const contribsWithEditables = contribList.filter(x => x.editable);
   const contribIdSet = new Set(contribsWithEditables.map(x => x.id));
   const [filteredSet, setFilteredSet] = useState(new Set(contribList.map(e => e.id)));
-  const [activeFilters, setActiveFilters] = useState({search: '', filters: []});
   const [selfAssignModalVisible, setSelfAssignModalVisible] = useState(false);
 
   const [checked, setChecked] = useState([]);
@@ -137,25 +137,33 @@ function EditableListDisplay({
           {value: 'ready_for_review', text: Translate.string('Ready for review')},
           {value: 'not_submitted', text: Translate.string('Not submitted')},
         ],
+        isMatch: (contrib, selectedOptions) =>
+          (contrib.c.editable && selectedOptions.includes(contrib.c.editable.state)) ||
+          (selectedOptions.includes('not_submitted') && !contrib.c.editable),
       },
       {
         key: 'editor',
         text: Translate.string('Editor'),
         options: _.uniqBy(
-          contribsWithEditables.map(c => c.editable.editor).filter(Boolean),
+          contribsWithEditables.map(c => c.editable.editor).filter(x => x),
           'id'
         ).map(({identifier, fullName}) => ({
           value: identifier,
           text: fullName,
         })),
+        isMatch: (contrib, selectedOptions) =>
+          contrib.c.editable &&
+          contrib.c.editable.editor &&
+          selectedOptions.includes(contrib.c.editable.editor.identifier),
       },
       {
         key: 'code',
         text: Translate.string('Program code'),
-        options: _.uniq(contribList.map(c => c.code).filter(Boolean)).map(code => ({
+        options: _.uniq(contribList.map(c => c.code).filter(x => x)).map(code => ({
           value: code,
           text: code,
         })),
+        isMatch: (contrib, selectedOptions) => selectedOptions.includes(contrib.c.code),
       },
       {
         key: 'keywords',
@@ -164,6 +172,8 @@ function EditableListDisplay({
           value: keyword,
           text: keyword,
         })),
+        isMatch: (contrib, selectedOptions) =>
+          contrib.c.keywords.some(k => selectedOptions.includes(k)),
       },
     ],
     [contribList, contribsWithEditables]
@@ -230,53 +240,21 @@ function EditableListDisplay({
     setSortedList(list => list.map(_mapper));
   };
 
-  const handleFilterChange = value => {
-    setActiveFilters(value);
-    const searchValue = value.search.toLowerCase().trim();
+  const filterableContribs = useMemo(
+    () =>
+      contribList.map(c => ({
+        id: c.id,
+        searchableId: c.friendlyId,
+        searchableFields:
+          c.editable && c.editable.editor
+            ? [c.title, c.code, c.editable.editor.fullName]
+            : [c.title, c.code],
+        c,
+      })),
+    [contribList]
+  );
 
-    const isMatch = result => {
-      if (searchValue) {
-        const match = searchValue.match(/^#(\d+)$/);
-        if (match) {
-          return result.friendlyId === +match[1];
-        }
-        if (
-          !(
-            result.title.toLowerCase().includes(searchValue) ||
-            (result.editable &&
-              result.editable.editor &&
-              result.editable.editor.fullName.toLowerCase().includes(searchValue)) ||
-            (result.code && result.code.toLowerCase().includes(searchValue))
-          )
-        ) {
-          return false;
-        }
-      }
-      const matchFilterType = (filterType, cond) => {
-        const filter = value.filters.find(f => f.filter.key === filterType);
-        return (
-          !filter || (cond ? cond(filter) : filter.selectedOptions.includes(result[filterType]))
-        );
-      };
-      return (
-        matchFilterType(
-          'editor',
-          f =>
-            result.editable &&
-            result.editable.editor &&
-            f.selectedOptions.includes(result.editable.editor.identifier)
-        ) &&
-        matchFilterType('code') &&
-        matchFilterType(
-          'state',
-          f =>
-            (result.editable && f.selectedOptions.includes(result.editable.state)) ||
-            (f.selectedOptions.includes('not_submitted') && !result.editable)
-        ) &&
-        matchFilterType('keywords', f => result.keywords.some(k => f.selectedOptions.includes(k)))
-      );
-    };
-    const filteredResults = new Set(contribList.filter(isMatch).map(e => e.id));
+  const handleFilterChange = filteredResults => {
     setFilteredSet(filteredResults);
     setSortedList(_sortList(sortBy, sortDirection, filteredResults));
     setChecked(
@@ -496,9 +474,9 @@ function EditableListDisplay({
             </>
           )}
         </div>
-        <EditableListFilter
+        <ListFilter
+          list={filterableContribs}
           filterOptions={filterOptions}
-          activeFilters={activeFilters}
           onChange={handleFilterChange}
         />
       </div>
@@ -609,151 +587,4 @@ EditableListDisplay.propTypes = {
   eventId: PropTypes.number.isRequired,
   management: PropTypes.bool.isRequired,
   canAssignSelf: PropTypes.bool.isRequired,
-};
-
-const filterShape = {
-  key: PropTypes.string.isRequired,
-  text: PropTypes.string.isRequired,
-  options: PropTypes.arrayOf(
-    PropTypes.shape({
-      value: PropTypes.string.isRequired,
-      text: PropTypes.string.isRequired,
-    })
-  ).isRequired,
-};
-
-function EditableListFilter({filterOptions, activeFilters, onChange}) {
-  const [openSubmenu, setOpenSubmenu] = useState(-1);
-
-  const handleSearchChange = search => onChange({search, filters: activeFilters.filters});
-  const handleFiltersChange = filters => onChange({search: activeFilters.search, filters});
-
-  const removeFilter = key => {
-    handleFiltersChange(activeFilters.filters.filter(f => f.filter.key !== key));
-  };
-
-  const toggleFilter = (filter, option) => {
-    const activeFilter = activeFilters.filters.find(f => f.filter.key === filter.key);
-    if (activeFilter) {
-      if (activeFilter.selectedOptions.length === 1 && activeFilter.selectedOptions[0] === option) {
-        removeFilter(filter.key);
-      } else {
-        const selectedOptions = activeFilter.selectedOptions.includes(option)
-          ? activeFilter.selectedOptions.filter(o => o !== option)
-          : [...activeFilter.selectedOptions, option];
-        handleFiltersChange([
-          ...activeFilters.filters.filter(f => f.filter.key !== filter.key),
-          {filter, selectedOptions},
-        ]);
-      }
-    } else {
-      handleFiltersChange([...activeFilters.filters, {filter, selectedOptions: [option]}]);
-    }
-  };
-
-  const isSelectedOption = (filterKey, option) =>
-    activeFilters.filters.find(f => f.filter.key === filterKey)?.selectedOptions.includes(option);
-
-  return (
-    <div>
-      <Dropdown
-        text={Translate.string('Filter')}
-        icon="filter"
-        direction="left"
-        labeled
-        button
-        className={activeFilters.filters.length > 0 ? 'orange icon' : 'icon'}
-        onClose={() => setOpenSubmenu(-1)}
-      >
-        <Dropdown.Menu styleName="filters-menu">
-          {activeFilters.filters.length > 0 ? (
-            <>
-              <div styleName="active-filters">
-                {_.sortBy(activeFilters.filters, 'filter.text').map(({filter, selectedOptions}) => (
-                  <Label key={filter.key} styleName="filter" className="fluid" color="orange" basic>
-                    <p>{filter.text}</p>
-                    <Label.Detail className="float right">
-                      {_.sortBy(
-                        filter.options.filter(o => selectedOptions.includes(o.value)),
-                        'text'
-                      ).map(o => (
-                        <p key={o.value}>{o.text}</p>
-                      ))}
-                    </Label.Detail>
-                    <Icon
-                      name="delete"
-                      styleName="right"
-                      onClick={evt => {
-                        evt.stopPropagation();
-                        setOpenSubmenu(-1);
-                        removeFilter(filter.key);
-                      }}
-                    />
-                  </Label>
-                ))}
-              </div>
-              <Dropdown.Item
-                text={Translate.string('Clear all filters')}
-                onClick={() => handleFiltersChange([])}
-              />
-            </>
-          ) : (
-            <Dropdown.Item text={Translate.string('No filters were added yet')} disabled />
-          )}
-          <Dropdown.Divider />
-          {_.sortBy(filterOptions, 'text').map(filter => (
-            <Dropdown
-              key={filter.key}
-              scrolling
-              icon={null}
-              className="item"
-              direction="right"
-              onOpen={() => setOpenSubmenu(filter.key)}
-              onBlur={evt => evt.stopPropagation()}
-              open={openSubmenu === filter.key}
-              disabled={filter.options.length === 0}
-              trigger={<Dropdown.Item text={filter.text} icon="plus" />}
-              options={_.sortBy(filter.options, 'text').map(({value, text}) => ({
-                key: value,
-                value,
-                text,
-                active: isSelectedOption(filter.key, value),
-                selected: false,
-                onClick: (evt, {value: v}) => toggleFilter(filter, v),
-              }))}
-            />
-          ))}
-        </Dropdown.Menu>
-      </Dropdown>
-      <Input
-        placeholder={Translate.string('Enter #id or search string')}
-        onChange={(evt, {value}) => handleSearchChange(value)}
-        value={activeFilters.search}
-        style={{width: '14em'}}
-        icon={
-          activeFilters.search
-            ? {
-                name: 'close',
-                link: true,
-                onClick: () => handleSearchChange(''),
-              }
-            : undefined
-        }
-      />
-    </div>
-  );
-}
-
-EditableListFilter.propTypes = {
-  filterOptions: PropTypes.arrayOf(PropTypes.shape(filterShape)).isRequired,
-  activeFilters: PropTypes.shape({
-    search: PropTypes.string.isRequired,
-    filters: PropTypes.arrayOf(
-      PropTypes.shape({
-        filter: PropTypes.shape(filterShape).isRequired,
-        selectedOptions: PropTypes.arrayOf(PropTypes.string).isRequired,
-      })
-    ).isRequired,
-  }).isRequired,
-  onChange: PropTypes.func.isRequired,
 };
