+++ indico/modules/events/editing/client/js/management/editable_type/EditableList.module.scss
@@ -15,44 +15,6 @@ $border: 1px solid $pastel-gray;
   justify-content: space-between;
 }
 
-.filters-menu {
-  max-width: 15em;
-}
-
-.active-filters {
-  display: flex;
-  flex-direction: column;
-  margin: 0.4em;
-  gap: 0.4em;
-}
-
-:global(.ui.label).filter {
-  display: flex;
-  margin: 0;
-
-  p {
-    margin-block-end: 0;
-  }
-
-  :global(.detail) {
-    overflow: hidden;
-
-    p {
-      overflow: hidden;
-      text-overflow: ellipsis;
-
-      &:not(:last-child) {
-        margin-block-end: 0.1em;
-      }
-    }
-  }
-
-  .right {
-    margin-left: auto;
-    padding-left: 1em;
-  }
-}
-
 .editable-list {
   .table {
     font-family: 'Muli', 'Helvetica Neue', Arial, Helvetica, sans-serif;
