+++ indico/web/client/js/react/components/index.js
@@ -29,6 +29,7 @@ export {default as StickyWithScrollBack} from './StickyWithScrollBack';
 export {default as EmailListField, FinalEmailList} from './EmailListField';
 export {default as PopoverDropdownMenu} from './PopoverDropdownMenu';
 export {default as ResponsivePopup} from './ResponsivePopup';
+export {default as ListFilter} from './ListFilter';
 export {default as MathJax} from './MathJax';
 export {default as FileSubmission} from './files/FileSubmission';
 export {default as FinalSingleFileManager} from './files/SingleFileManager';
